+++ src/store.js
@@ -102,7 +102,13 @@ import { shareModalReducers } from 'connectors/share-modal/share-modal.state'
 import { listenReducers } from 'connectors/listen/listen.state'
 import { listenSagas } from 'connectors/listen/listen.state'
 
+import { readerSettingsReducers } from 'containers/read/reader-settings.state'
+import { readerSettingsSagas } from 'containers/read/reader-settings.state'
+
 //Items (From the graph)
+import { readerReducers } from 'containers/read/reader.state'
+import { readerSagas } from 'containers/read/reader.state'
+
 import { itemsReducers } from 'connectors/items/items.state'
 import { itemsSavedReducers } from 'connectors/items/items-saved.state'
 import { itemsSavedSagas } from 'connectors/items/items-saved.state'
@@ -121,6 +127,9 @@ import { mutationTaggingSagas } from 'connectors/items/mutation-tagging.state'
 
 import { mutationUpsertSagas } from 'connectors/items/mutation-upsert.state'
 
+import { mutationHighlightReducers } from 'connectors/items/mutation-highlight.state'
+import { mutationHighlightSagas } from 'connectors/items/mutation-highlight.state'
+
 import { mutationBulkReducers } from 'connectors/items/mutations-bulk.state'
 import { mutationBulkSagas } from 'connectors/items/mutations-bulk.state'
 
@@ -138,6 +147,7 @@ const itemReducers = {
   mutationArchive: mutationArchiveReducers,
   mutationDelete: mutationDeleteReducers,
   mutationTagging: mutationTaggingReducers,
+  mutationHighlight: mutationHighlightReducers,
   listen: listenReducers
 }
 
@@ -182,8 +192,10 @@ const libraryReducers = {
   profileItemsByIds: profileItemsReducers
 }
 
-const readerReducers = {
-  reader: readReducers
+const readerViewReducers = {
+  reader: readReducers,
+  readerSettings: readerSettingsReducers,
+  readerGraph: readerReducers
 }
 
 const marketingReducers = {
@@ -216,7 +228,7 @@ export const rootReducer = combineReducers({
   ...discoverReducers,
   ...collectionReducer,
   ...libraryReducers,
-  ...readerReducers,
+  ...readerViewReducers,
   ...userAccountReducers,
   home: homeReducers,
   homeItemsById: homeItemsReducers,
@@ -249,6 +261,8 @@ function* rootSaga() {
     ...itemShareSagas,
     ...itemReportSagas,
     ...readSagas,
+    ...readerSettingsSagas,
+    ...readerSagas, //graph
     ...homeSagas,
     ...homeItemsSagas,
     ...userMessageSagas,
@@ -265,7 +279,8 @@ function* rootSaga() {
     ...mutationUpsertSagas,
     ...mutationBulkSagas,
     ...mutationTaggingSagas,
-    ...listenSagas
+    ...mutationHighlightSagas,
+    ...listenSagas    
   ])
 }